--- deepstream/pravega_protocol_adapter/Makefile.test
@@ -1,37 +1,0 @@
-################################################################################
-# Copyright (c) 2018-2020, NVIDIA CORPORATION. All rights reserved.
-#
-# NVIDIA Corporation and its licensors retain all intellectual property
-# and proprietary rights in and to this software, related documentation
-# and any modifications thereto.  Any use, reproduction, disclosure or
-# distribution of this software and related documentation without an express
-# license agreement from NVIDIA Corporation is strictly prohibited.
-#
-################################################################################
-# this  Makefile is to be used to build the test application to exercise the sample_proto protocol adaptor
-CXX:=g++
-NVDS_VERSION:=5.1
-DS_INC:=/opt/nvidia/deepstream/deepstream-$(NVDS_VERSION)/sources/includes
-DS_LIB:=/opt/nvidia/deepstream/deepstream-$(NVDS_VERSION)/lib
-
-SYNC_SEND_BIN:=target/test_pravega_proto_sync
-ASYNC_SEND_BIN:=target/test_pravega_proto_async
-
-SYNC_SEND_SRCS:=test_pravega_proto_sync.cpp
-ASYNC_SEND_SRCS:=test_pravega_proto_async.cpp
-
-CXXFLAGS:= -I$(DS_INC) -rdynamic
-LDFLAGS:= -L$(DS_LIB) -ldl -Wl,-rpath=$(DS_LIB)
-
-default: all
-
-all: $(SYNC_SEND_BIN) $(ASYNC_SEND_BIN)
-
-$(SYNC_SEND_BIN) : $(SYNC_SEND_SRCS)
-	$(CXX) -o $@ $^  $(CXXFLAGS) $(LDFLAGS)
-
-$(ASYNC_SEND_BIN) : $(ASYNC_SEND_SRCS)
-	$(CXX) -o $@ $^  $(CXXFLAGS) $(LDFLAGS)
-
-clean:
-	rm -rf $(SYNC_SEND_BIN) $(ASYNC_SEND_BIN)
